Clean up parsing and add simple_step API
