+++ src/bang/__init__.py
@@ -1,3 +1,3 @@
-from .core import PBN, load_from_file, german_gpu_run
+from .core import PBN, load_from_file
 
-__all__ = ['PBN', 'load_from_file', 'german_gpu_run']
+__all__ = ['PBN', 'load_from_file']
