+++ src/bang/core/__init__.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-# from ._gpu_stable import german_gpu_run, initialise_PBN
-from .PBN import PBN, load_from_file
-from .simulation import german_gpu_run
+from bang.core.PBN import PBN, load_from_file
 
-__all__ = ["PBN", "load_from_file", "german_gpu_run"]
+__all__ = ["PBN", "load_from_file"]
