+++ tests/PBN_test2.py
@@ -1,5 +1,7 @@
 import bang
 
 pbn = bang.load_from_file("tests/test2.pbn", "assa")
-bang.initialise_PBN(pbn)
-bang.german_gpu_run()
+
+print(pbn)
+
+pbn.simple_steps(1000)
